backend/requirements.txt
fastapi
sqlalchemy
psycopg2-binary
boto3
uvicorn[standard]
ffmpeg-python
python-multipart
